"9:10am. I am up. Let me chill for a while and then I will start.
9:30am. Let me start. Today my goal will be to watch 10 lectures from the algebraic calculus playlist. I do not want to dwell too long on this. It was not my intention to watch like 1 per day.

Yesterday, I finally used a proof assistant to help me do what I was messing up on paper, so I am quite satisfied. It is the first time I had this experience of using the computer not just to calculate, but to directly reason about math. Now that I've broken the ice with this kind of proving style and I am a step closer to being complete as a programmer, I'll leave it aside for a while until it is time to get back into the PRML book again.

What I did yesterday was just so right, but I need to focus on what is in front of me.

Actually let me linger for a bit longer to highlight what I have done.

I've ensured that I will never make any of those dumb rewriting errors ever again. This is just so enormous as I've eliminated the worst class of math errors for good. Had I been taught to do math this way from the start rather than with pen and paper, just what could I have accomplished by now? Truly my life has been a waste.

I've wasted a lot of time, but if I can stop being human, then the time spent here will just be a drop in the ocean. I've always been aware of this, but the trials continue to chip away at my will. The power and the Inspired Desire will heal me once I get them.

9:45am. Now let me focus. Lecture watching time start!"