"9am. I am up. Let me chill for a while and then I will start.
https://www.reddit.com/r/MachineLearning/comments/d01tq3/discussion_categorytheoretic_approach_to_machine/

I noticed this thread last night. I think I will finally take a look at it today.

9:35am. Let me start. I'll check out the thread later.

It is time to watch some lectures.

I really hope the pace picks up from here.

10:05am. https://youtu.be/bPvinbhxmU0?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1424

It is not obvious to me where `(n-1)n(n+1)` is coming from. It might be worth figuring out how to do this derivation in Agda. Since the thing is recursive, it is necessary to infer the endpoint ahead of time, but I should be able to do this on my own. Later. Forget it for now.

10:20am. https://youtu.be/VtlOaQy2DfU?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Done with the first one. Yeah, when I am done with all of this, I am going to have to sit down and do some math. I am thinking about how Bayesian inference could be done again. Well, it could not be done, but I need to clarify some things regardless.

No forget it. This kind of thinking never goes anywhere. I need a real insight.

My thinking right now is just too loose.

10:25am. Focus, let me watch the second lecture.

10:55am. https://www.youtube.com/watch?v=rQzfHbkltkY&list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&index=14

Now come the third. My focus is much better in the morning as is usual and I do not feel as bored as I no doubt will in a few hours.

https://youtu.be/rQzfHbkltkY?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=97

Hmmm, the arithmetic series was not by Gauss? I've always thought that story was too good to be true.

11:25am. I am getting hit by inspiration again.

I've realized something. I am thinking about NNs from a Bayesian perspective and could just see for the first time that necessarily the weights in the upper layers do not necessarily have to depend on the weights in the previous ones.

Yesterday I had that insight of how probability distributions can be combined - they can simply be added up.

The problem of weight dependencies can be simplified by turning the dependencies to the inputs. Usually this would not be possible, but the inputs can be turned into a probability distribution by integrating them from the previous layer.

11:30am. Actually this is not completely true, as doing this would entail loss of information and introduce optimization dynamics. The thing would no longer be Bayesian, but an approximation of it. And the input even a distribution would still be too big.

I've really gotten to thinking of CFR as a Bayesian inference algorithm which it really is. It is not like Q learning. It does not propagate the rewards backwards through time. Rather it does filtering much like a genetic algorithm would. I guess this does explain its outperformance. Assuming all the models up front and doing filtering is definitely better than assuming a single model and then doing credit assignment.

11:40am. I keep trying to talk myself into going back into the lecture, but the surge of inspiration is going.

I've realized something else just now. If I turn inputs into a probability distribution, I can in fact have it backward channels in order to do credit assignment.

I do not have to just add the probability masses together. In a very real sense, if I keep track of the additions, I can reverse them back to the weights which produced them.

1:40pm. Yeah, this CFR style of backpropagating rewards would work for anything, not just tabular games.

One thing I've realized though - the idea sucks for anything other than tabular agents.

When you get a gust of inspiration like this, it seems like you've solved something important, but as a result of these thought experiments I've finally realized how much sampling sucks.

It sucks massively.

If you were doing all the reweightings in parallel, then you get optimal credit assignment.

Change that to sampling, and you get very poor credit assignment. Hierarchies won't change much here.

I can't imagine this beating actual backprop when it comes to NNs.

1:45pm. It cannot be naive sampling. It has to be something that manifests an instance from the shared posterior.

1:50pm. Ah, I do not get it at all. The Bayesian way is the most fundamental of all. But if so, where does that leave Q learning style algorithms. Seen from the perspective of Bayesian inference, I do not understand Q learning at all.

Is it right or wrong? Both CFR and Q learning cannot both be right.

1:55pm. I am just mentally exhausting myself with this. Let me get back to the lectures.

One thing that is certain from all of this is that I am focus too much on learning the task. A real learning system would focus much more on itself. In that respect, what I have written in the monthly update is perfect. That perspective if I can elaborate it will lead me to the solution."