"2:25pm. The CFR algorithm is really coming back vividly to me now. Besides backproping credit probabilistically, it had all sorts of things like inplace updates and opponent probabilities taken into consideration. I forgot about those.
There is more to Bayesian inference than it meets the eye. I have the general framework of it now, but that is still not quite enough to get to CFR from where I am. There are things still missing in my mind.

I need to figure out the link between BI and optimization. This course is exactly what I need in that respect. I'll eventually tie it together with abstract algebra. In fact, I'll need to bring all the science up to this point together in order to come to a conclusion.

I'll do it bit by bit.

I am not slacking at all, but somehow I've yet to watch even 3 lectures. I do not think I'll be doing 10 today either. But let me see if I can manage 5 like yesterday.

2:55pm. https://youtu.be/jx_JR5xD9Ko?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=275

Thunder. I might as well take a nap now.

5:40pm. This sucks. Let me see if I can get one more lecture done for the day. That will bring my total to 4.

5:50pm. Done with lunch. Let me get this out of the way.

6:10pm. https://youtu.be/jx_JR5xD9Ko?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1173

I can't follow this anymore. I'll stop here. That burst of inspiration and then brooding in bed over my math inadequacies have sapped my energy.

The way everything is going for me is just the worst. I picked the low hanging fruit with Spiral.

Everything that is left is beyond my talent. It does not mean it can't be done, but to do it I am going to have to go through hell...and not the good kind of hell.

Today, I finally understood that I should not interpret sampling as something that will get me the right answer asymptotically, but something that is worse the more samples are needed to get full coverage. I admit when I first saw Monte Carlo techniques and heard that it gets accurate results for any probability distribution I admit, the image I formed then and the image I have now after some thinking about it is not at all the same.

These bursts of inspiration today and yesterday are building on each other, but it is just making me realize how much I need to avoid the hacky methods in favor of saying something intelligent about learning.

What is the best dimensionality reduction technique? It is not SVD, GANs or anything stupid like that, but programs themselves.

I've worked hard for the past 4.5 years to get good at things the computers should be doing themselves.

I should have just made up my mind and went for it at the end of elementary school. I should have just aimed for AI from the start rather than hedging my bets endlessly.

This is going to be the last thing I do. I will learn all the math that is needed, learn all there is about programming and create those agents. If I cannot do it before my life ends, then that life will have been worthless.

But right now I am still alive, and I can still keep fighting. The universe might be torturous in the way it keeps throwing too difficult challenges, but it is not outright murderous.

6:20pm. For now let me rest. Hopefully, I won't be interrupted by dumb surges of either inspiration or electricity."