"10:05am. I seem to be getting more and more into Long War as time goes on. I do not know how I'd ever stop playing without the crashes to remind me that it is time to quit.
10:15am. Let me start the lectures. I am usually into them in the morning thankfully enough.

10:40am. https://youtu.be/qmMs6tf8qZ8?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Let me move on to this one. No way am I going to be able to remember what came in the previous lecture.

But it does not feel like any of the individual steps are hard.

11:15am. https://youtu.be/qmMs6tf8qZ8?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1413

Watching this stuff about matrices and their inverses is making me realize that I do not understand a very important thing when it comes to NNs.

https://en.wikipedia.org/wiki/Bayesian_linear_regression

Agh, I can't find the formula I want, but I remember Bayesian regression being `(X + lambda * I)^-1 b`. Meaning it is just like linear regression, but with some regularization.

What I want to know is how this would work for two layers.

I imagine you'd first generate some data using random networks, but then the issue is that the last layer only has the targets, but the first layer only has the inputs.

Given a bunch of targets, what would be the optimal inputs for that last layer?

I do not understand it. There is just so much I do not understand.

I really like algebraic calculus so far as it tries to bring linear algebra together with calculus in a modern, computable fashion. What I will need to do once I start studying topology is bring that together with type theory.

11:25am. Hmmmm...I keep remembering the claim by Wolfgang that the brain generates noise. What if the inputs to the last layer were all just noise? I could generate a bunch of noise targets and have the first layer solve for that...

11:30am. No then I am not doing hierarchical learning in any fashion, but just unsupervised learning. It would be extremely limited.

Forget this idea. Let me just move on with the lecture. I do not want to waste hours brainstorming. I need to get into the studies.

11:40am. https://youtu.be/E80wXXy09S4?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Finally, it is time for this one. I have no idea how I only watched 1.5 lectures, but let me stop here so I can eat.

The last two lectures had a lot of stuff in them, and I won't be able to remember this until I take time to practice it which I won't. But I am really not interested in sums of polynomials. I want to see what will come out of this course later.

I'll always be able to revisit this if I really need it. This is the beauty of self education."