"1:10pm. It is time to resume. For the next four hours, I will just watch lectures.
1:15pm. https://youtu.be/E80wXXy09S4?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=255

Given that a multiset is unordered, how can you associate a list to it...

...Ah, I get it. It it associating a list to the numbers themselves. The positions only coincided in this case with the example.

1:30pm. https://youtu.be/E80wXXy09S4?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=681

Egh, rationals do not really work with unordered lists. This kind of mental trickery I hate in math.

1:55pm. https://youtu.be/r26fH4h2Wm4?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Finally on the 4th lecture, and I can feel the boredom starting to set in. Let me take a break here.

2:55pm. https://youtu.be/A9-fT-QTRH8?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Finally on lecture 19. I really hope that this course picks up and starts talking about actual uses, but I guess this is how much one could expect from a theory course.

I took this challenge and I am going to go through with it. At this pace, even if I do only 5 lectures per day, I will still finish the course in a week.

Let me take a short break. Now I am wondering why can't I find this thing on OpenLearning. I wanted to look for the part 2 there, but I can even find the first part.

https://www.openlearning.com/courses/algebraic-calculus-one/

Well, here it is...

For some reason it does not show up on search.

I am starting to hear thunder in the distance. Damn it.

Any more and I am going to have to run.

3pm. There is nothing on part two. Even part one is going to start in 5 months.

3:05pm. http://www.wildegg.com/youtube-differential-geometry.html

I am going to stop here due to weather again.

When I am done with calculus and topology, I think I am going to go through this course properly even though it made me fall asleep last time during my information geometry days.

4pm. The thing blew by. Let me resume. I still have steam for two more lectures. Let me do it.

4:35pm. https://youtu.be/6fyEoWN7vCM?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

One more lecture and then I think I will stop. I guess 10 lecture per day is way too much for me. I just do not have the focus for something like this which is so beyond my talent.

4:40pm. Ok, focus me. Let me just do this.

5:10pm. https://youtu.be/cv8kDOaenpU?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Done. Finally done for the day. The first 3 videos were palatable, but the last two are torture.

5:15pm. Oh, did the English version finally go past the scanlations. Yeah, the last volume me was 8. Nice. Downloading...

Reading this, Long War, bath, lunch, watching Symphogear is what I am going to be doing for the rest of the day. I had enough of math. For a person like me this is pure torture.

Unlike programming, I do not have an intrinsic interest for math. This is really the worst possible situation for me. The rest of the field is worthless, the pieces are not coming together, and to actually make something myself, I am going to have to study like crazy.

I am too prideful to say that I am just going to be able to make it work out somehow. It is obvious to me that I do not have the talent to push through this hurdle, and yet, this is important enough that I cannot let it lie. It is like the alien invasion. Regardless of your physical fitness and training, you just have join the Resistance and fight. Otherwise the Avatar Project will be completed and everybody will die.

It is absolutely the worst to fight knowing only failure awaits, but that is the grim mindset that I've had to accept when I finally decided to go down this route.

5:25pm. If I only had this talent, my life right now would be so much easier. I guess NEETdom is my destiny. I do not think that the average NEETs actually study math or programming like me, but unless I can manage something, my impact on the world will be the same.

I'll let the next few years decide it. I'll push myself to the very limit, but if I absolutely can't make an iota of progress despite learning all the math that I can and mastering homotopy type theory, then I am just going to accept that the talent is the deciding factor and do something easier like actually programming for money. That would be the sane thing to do.

I did trading for six years and eventually threw in the towel on that.

But probably, even though I say this, I will just keep going for as long as I can. The real me would do this for decades even if he cannot hack it.

Aren't those kinds of people the ones that I admire?"