"10:20am. I got up late today. Let me chill for more minutes and then I will start the lectures.
11:20am. https://youtu.be/ZlE36x7cXqM?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1443

Damn, that is some nasty thunder out of the blue. The weather has been the worst for the past few days. I guess the summer is really over.

11:45am. Let me resume. The thing blew over.

11:55am. https://youtu.be/P9XYzST6SZU?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

I hate it when I get interrupted like that. Let me go for one more video.

...I really hate it that there is no material on machine learning that just explains all the aspects of intelligence in a clear and easy to understand fashion. This physics and math stuff is like a kid's toy compared to what happens in ML. I am going to have to go through hell to get anything of intelligence.

This kind of work is not glorious at all.

But I will do it.

12:30pm. https://youtu.be/-VAf0GPhJXc?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Let me stop here. 24/60...

Hmmm, wasn't it 59 yesterday? Are new videos getting added?

Ah, it seems that the videos past 58 are reworkings of earlier videos.

Ok, I am safe. I only need to watch 58 then.

6 more I will be past the halfway point."