"1:30pm. Let me chill for a while more and then I will start. I actually quite like the first ep of Machikado Mazoku. I think I'll give rest a try.
2pm. Enough of this. Let me resume the lectures.

2:05pm. I have a bit of inspiration. I am starting to see, or perhaps finally accepting, that to do ML I am going to have to study physics. Physics has so much to tell about how to optimize ML models. It is probably the single best source of inspiration out there. Yet, I've completely ignored it in my education.

2:10pm. I need more, more, more...

The pieces are all there. I just need to reach further and put them together.

The true language is beyond my ability, but the basics of learning is something I should be able to grasp. There is a feeling in me.

Just like when I was agitated by F# over not being able to do the ML library, right now I just can't believe this is the limit of my ability. It should be possible to do more. Futilely grasping in the darkness like I am doing now just cannot be the end of this story.

2:25pm. https://youtu.be/-VAf0GPhJXc?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=459

I am a bit confused as to what he is drawing here.

https://en.wikipedia.org/wiki/Pole_and_polar

Let me read this.

3:05pm. Rather than watching the lecture I am getting consumed by inspiration again.

The notion of physical models is just so powerful. If physical models were programs just how would they be?

And conversely, if the learning models were physical models, what would their properties be? And what should their properties not be?

The analogy is not that crazy. The Coq Software Foundations proofs are an aberration. The way I prove correctness of programs in my head is not like that. It is more like a models controlled by another models. It is just programs all the way up, but not the standard kinds of programs.

The standard kinds of programs are narrowly focused due to having to be expressed explicitly. Programs as learning models would be different.

3:10pm. There is one thing I am sure. The conception of inference or optimization is just too narrow to talk about properties of programs. I need to widen my view in order to get the full breadth of it. Optimization is secondary.

Why not think of RL agents as multidimensional objects going through some space. Then rewards could just be vectors that push it this way or the other.

What I did in the monthly update is make a prediction. Saying that paths have to play a role would in physical terms saying that rather than it being like an object flying through space, it is more like a fabric being stretched. It posits an altogether different property.

3:15pm. The mental tools to understand learning seen purely through the lens of optimization are not enough. I need to think about movement more generally. And movement at it most general is computation.

All the programs I've ever written were points.

I need to think what truly multidimensional programs would look like. Let me see if I can focus on the lectures."