"3:30pm. https://youtu.be/-VAf0GPhJXc?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1323
This has been interesting so far even if boredom is starting to kick in for me at this point. Yeah, using chords does make sense and checks out algebraically. Even though when `t = u` it would violate that earlier division...

"I don't agree that you've avoided the issue of limits. Your derivation technique is exactly the same as the conventional one for d/dx (x^2). The problem is that you divided by (t-u) assuming that t is not equal to u, and then set t = u. Without justification, you can't assume that's a valid thing to do. Providing that justification is part of what the theory of limits is for."

"I think part of the difficulty is in deciding what roles t and u play. If they are numbers then your objection is reasonable, however if they are variables then we are just performing algebra."

One of the commenters does bring it up. Actually I agree that he did not avoid the issue with limits. First he is assuming that t is not u, and then he is assuming that is what they are. That has to be some kind of contradiction.

"To cancel factors of (t - u) when they are equal means a case of "division by zero". I'm not sure that you can do this cancelling in algebra, first assuming the variables are different without taking care of that case (t = u) in the first place. In calculus this is called a removable discontinuity (the division by zero point), you can work with values bigger and smaller than that zero value, while the discontinuity point is treated separately."

"One doesn't need to insist that 't' and 'u' should be numbers. They are polynumbers. When you plug in values for 't' and 'u' it can turn out that t-u becomes 0, but even after that fact I want to insist that t-u is nonzero as a bipolynumber. In the mset framework we may define t=[[1]] and u=[[2]] and then t≠u unambiguously."

...Eagh, I have no idea. Let me just watch the lecture.

This one and one more and I think I'll be done for the day.

3:55pm. Focus me. Just watch the lecture. The physical model analogies are not going to get me far today.

I need to laser in on the task at hand.

4:25pm. https://youtu.be/GkXd6VXQM-I?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1033

I need to stop here for a bit. I can't focus on this anymore at all. With this course it seems that 5 lectures per day is firmly my limit.

4:50pm. Let me resume. I only have half of that lecture left. Let me backtrack a little and I will finish it.

5:05pm. https://youtu.be/GkXd6VXQM-I?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1754

Yeah, I finally remember this from differential geometry.

5:15pm. https://youtu.be/XdyrzOyrtwc?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Let me go on for a bit more. I am dead tired, but I want to push for a bit further.

6pm. https://youtu.be/XdyrzOyrtwc?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=2018

Ugh, thunder. Let me stop here for the day."