"9:50am. Seriously, fuck weather. I've been in bed for over 15h now and it is suddenly freezing in my room.
10:05am. The summer really is over. There is usually a day to highlight that.

Despite being in bed for so long, I had trouble sleeping so I am groggy as hell right now.

Before I resume slacking let me talk about reversibility for a bit since I had some inspiration over the night. I am going to find a way to fit what I am going to say now into the next monthly update.

10:25am. ...Well, I do not feel like writing about it right now. I am drained.

10:40am. Let me start the lectures. I'll finish the one from yesterday and watch two more.

10:55am. I am so tired. I can't focus a thing. This is what going to bed at 6pm, but falling asleep at 2am will do to you.

https://youtu.be/5RTbX1bz0Zg?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

No forget it. This is a futile cause today. I might as well go to bed.

11am. Since 11am is too early to go to bed, let me talk about that idea I had during the night.

Well, it is quite simple.

---

Let me compare reversibility to backprop.

Consider the function `f x = x*x`. Suppose you have 3 targets `1,000,000`, `4,000,000`, `9,000,000`.

What would gradient descent do with squared error `1/2 * (target - x^2)^2`? The `cost` would be `target - x^2` as per the derivative of the squared error function. And the update would be `cost * 2 * x`. The `2 * x` is because the derivative of `x*x` is that. So if you started `x` out at 1, the update for the `1,000,000` target would be `2,000,000`. That is a large number.

With reversibility, you'd just go the opposite way. So the reverse of `1,000,000`, `4,000,000`, `9,000,000` would be `1,000`, `2,000` and `3,000` respectively just by taking the square root of the targets. Suppose you started at `0` and were doing iterative reversible updates, then you'd have those values as the analogues of the gradient.

Starting at zero with backprop for this function would never update at all due to the `2 * x` term. And suppose you started at something like `2,000` and got the `9,000,000` as target. Then the update would be `(9,000,000 - 4,000,000) * 2 * 2,000`.

`20,000,000,000` would be the update.

How would you even set the learning rate for a function like this?

I am using this example to highlight how backpropagation does not even remotely do the sensible thing even in the 1D case for a very simple function. Forget reinforcement learning, it cannot even do the square properly.

---

11:15am. That having said, it is not obvious to me how to scale reversibility past a single variable. When you have `a + b` rather than just `x` reversing that would turn into an optimization problem.

It is not obvious at all how to reverse a model across multiple example. The above seems to indicate that there is some sort of optimization going on as `x` cannot be `1,000`, `2,000` and `3,000` at the same time. One would probably have to pick `2,000` somehow, but how to get that is a mystery.

11:20am. Isn't it a lucky coincidence that at around this time a major breakthrough in type theory is all about reversibility?

I cannot leave this strike of fortune aside. It definitely has to be some kind of omen that I should go in this direction. Plus, usually useless fields like topology tend to spring to life in times like these.

11:30am. Now if only my body would listen to me.

Damn, I am so tired. I do not feel like even games or manga."