" 12:30pm. While I am tired, I can't just waste the whole day doing nothing. I am going to grit my teeth and just watch a few lectures.
1:20pm. Done with the first lecture. Yeah, I am dead tired. Let me just take a slight break and then it is time to do the next one.

1:25pm. Ok, let me do it. My focus is barely half of what it is usually, but I will pull through this.

https://youtu.be/-xHkn1po6fo?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=190

Hmmm...I forgot the details, but wasn't this covered in the previous lectures? Why is he going at it again from a different angle?

2:05pm. https://youtu.be/3BLcDGUKiPs?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Time for a break again.

2:45pm. 40 minutes went by already? Really? Damn. Ok, as soon as the song finishes, I am going right into the next video.

2:55pm. https://youtu.be/3BLcDGUKiPs?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=442

I know I said I would keep going, but my focus is really rock bottom right now. Even on a good day I can barely manage to do 5. Today maybe I should just be happy that I managed 3.

I could technically keep forcing myself to watch this, but I am really not paying proper attention to this at this point. My brain itself is rebelling as I try to ingest this material.

3:10pm. Yeah, I am done for the day. This sucks.

Usually, you'd expect that going to bed early would result in more rest, but that is not necessarily the case for me. Well, at least, when I sleep badly one night this usually preps me for some great sleep the following night.

3:15pm. Screw it. Just for today, I am going to stop a few hours earlier than usual and spend that time on fun. I'll waste the time left.

It is strange. When I am in the thrall of my obsession even though I should be resting, my brain is always churning away at it on the side. So it won't be strict laziness as I do not have the capacity to allow myself this. But I just do not understand why this obsession unlike with programming does not result in positive feedback when I study math.

Is my obsession just too weak compared to what it could be?

...

Hmmm...well suppose programming is math, just of a different kind. Then maybe there are things I like and dislike. Like, I cannot stand real analysis at all, but calculus is a bit better. And I've liked what I've seen of topology so far.

Maybe topology turn out to be up my alley?

Formal theorem proving with computers while tough, was not too bad. The game was fun enough that I could do it for weeks without end.

These calculus lectures on the other hand are really killing me. I really cannot take much more of this. Thankfully I will be done in less than a week and will be able to move on to topology.

3:40pm. Ok, enough of this. Enough 4chan on the side. Games. Manga. Anime. Go!"