"8:15am. Ah, I slept great. As expected, bad night are follow by good ones for me. Wonderful. Let me chill for a bit and then I will start with the lectures.
8:40am. https://pron.github.io/posts/computation-logic-algebra-pt2

Come to think of it, I never finished this series. Let me do so, read the Satanophany raw and then I will get through my quota of lectures for the day. Yesterday's Long War session was most enjoyable. I did a supply raid without infiltration and it has kept me on the edge of my seat the whole time. Because of the sudden drop in temperature it is very cold in my room, but the tension of the game made me sweat.

9:10am. "Neither the the hand nor the intellect do much on their own; the work is no less that of the intellect than of the hand. And as the mind guides the motion of the hand, the hand protects the understsanding of the mind"

`understsanding` should be `understanding`. This is a great quote by the way.

9:35am. https://pron.github.io/posts/computation-logic-algebra-pt3
"An almost metaphysical system of abstract signs, by which the motion of the hand performs the office of the mind."

I need to quote this again as it is such a great quote.

Done with part 2. I want to mail the correction to pron for that typo, but I'd rather not use twitter for that.

Let me start the lectures. By the way, today's Satanophany chapter was glorious.

9:40am. Enough wasting time. Let me get this thing rolling.

9:55am. https://youtu.be/3BLcDGUKiPs?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=598

Ahhh, I see. This gives me flashbacks to my school days. I think there was a programming problem I never got done because I did now know firmly how to determine whether a point is to the side of some line. I definitely encountered such a thing, but the details of the problem have nearly completely faded to memory by now.

Yesterday I was really confused why he was bringing up the sides all of a sudden, but now I see what he is getting at. Whether the signed areas are positive or negative is what determines the side.

10:05am. Actually, now that I am thinking about this more closely, are areas really the right thing here?

https://youtu.be/3BLcDGUKiPs?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=423

Looking at this image again you can have the points on the opposite sides without crossing.

But then you could calculate more areas in order to determine whether they are actually crossing.

Still, wasn't there a meet of two lines formula...ah, no wait. Lines are infinite. Here the problem is the one of dealing with meets of intervals.

10:25am. I am having some inspiration, so I just had to pause a bit to think.

Let me add a few notes to that examples.

[Edit: Consider that if you solve for `x` using reversibility, since you are no longer using squared error, the most sensible thing to set it to would be `2,000`. When squared to `4,000,000`, this is not the mean (which square error calculates) but the median. This is not bad as the property of not putting too much weight on outliers is a good one for an algorithm to have.

Still, it is not obvious how to go from median to the mean. If the solutions turn out to be exact then it does not matter either way, but when approximations must be made like here then careful consideration of the structure of the model must be made. With reversibility, the cost function is no longer explicit like in the gradient based methods, but seems to depend on the details on the model.

I am wondering if the phenomenon is linked to [hyperbolic vs exponential discounting](https://agentmodels.org/chapters/5a-time-inconsistency.html) of credit as found in biological organisms.]

This is what I put in. Let me resume the lectures.

10:55am. https://youtu.be/BVNW7cssNA8?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Finally done with the first lecture.

I am thinking that to get reversibility of multiplication, maybe it would not be a bad idea to get rid of the zero entirely from the algebra. I mean it is not like there is some need for the hidden layers to have this. It might be possible to go further and to make an algebra where there are no singularities for matrix inversion. Still it is a mystery to me how I would handle non-square matrices in that case.

Well, let me just keep going.

11am. https://youtu.be/BVNW7cssNA8?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=188

Hmmm, it really wasn't always obvious that a cannonball goes in a parabola when launched? Mindblowing. I wonder what beliefs of today's era will get utterly discredited.

11:20am. I need to focus on the lectures, not reading the LW2 posts.

11:45am. https://www.youtube.com/watch?v=mpy_WM_s0eE&list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&index=31

Let me stop here. I am definitely going to fill my quota of 5 lectures today. No doubt about it.

12:40pm. "… We suppose, it would seem, that concepts sprout in the individual mind like leaves on a tree, and we think to discover their nature by studying their birth: we seek to define them psychologically, in terms of the nature of the human mind. But this account makes everything subjective, and if we follow it through to the end, does away with truth. What is known as the history of concepts is really a history either of our knowledge of concepts or of the meanings of words. Often it is only after immense intellectual effort, which may have continued over centuries, that humanity at last succeeds in achieving knowledge of a concept in its pure form, in stripping off the irrelevant accretions which veil it from the eyes of the mind. What, then, are we to say of those who, instead of advancing this work where it is not yet completed, despise it, and betake themselves to the nursery, or bury themselves in the remotest conceivable periods of human evolution, there to discover, like John Stuart Mill, some gingerbread or pebble arithmetic! It remains only to ascribe to the flavour of the bread some special meaning for the concept of number. A procedure like this is surely the very reverse of rational, and as unmathematical, at any rate, as it could well be. No wonder the mathematicians turn their backs on it. Do the concepts, as we approach their supposed sources, reveal themselves in peculiar purity?"

This is really a great quote.

Note by pron: "rege may be criticizing the very goal of my text, but sometimes, after the concepts have been purified so much that their origins are forgotten, traveling back to their source — like tracing the evolution of the eye — may reveal exactly which stages of development were most crucial to which aspect of the concept, and so may reveal some possible paths forward. Not at all; we see everything as through a fog, blurred and undifferentiated."

12:40pm.

---

… To quote Mill: “The doctrine that we can discover facts, detect the hidden processes of nature, by an artful manipulation of language, is so contrary to common sense, that a person must have made some advances in philosophy to believe it.

---

This is obviously wrong on the account of past few millenia of progress, but it does have intuitive appeal.

1:55pm. Let me resume. Time to get the other 3 out of the way."