"3:25pm. I am getting distracted by that annoying back and forth, so let me get my focus back and finally finish the third lecture.
3:35pm. No actually, I am super pissed right now. I can't believe I haven't even watched 3 lectures by now. But this is the usual problems for me. What I am doing here is beyond my talent so I get knocked off balance much more easily that I would had it been regular programming for example.

Fight me! Reality might be bullying me, but I will persist!

4:05pm. https://youtu.be/EIDWc5lgt_M?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Done with the lecture. Time for the 4th one. I have my focus back. This is not that uninteresting, but there is a lot to take in.

4:25pm. https://youtu.be/EIDWc5lgt_M?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=624

Hmmm...this is interesting. Pythagoeran theorem can be proven? Amazing. I actually had no idea it could be proven like this.

4:45pm. That annoyance distracted me again. I'll manage to get through this somehow. 10m left in this lectures, and then another one and then I am done for the day.

Because of math, my stress is already high at this point so I really do need to unwind. I can't allow myself to become obsessive about potential replies.

5:20pm. https://youtu.be/EIDWc5lgt_M?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1572

I do not understand why that middle formula for the area is valid. Why would it give equal result to the first one?

Ahh, I see. It does not really matter what those other two sides are. I remember this from school now. It is interesting what mistakes even geniuses make.

5:30pm. Ahhhh...I've had an important insight into how to do the inverse even when the matrix is nonsquare. It can be padded!

Still while that might improve the conditioning and make everything reversible, I am not sure what the theory of that would be.

If I think about it, it is quite a miracle that a matrix can be inverted in the first place. Considering the columns are isolated in the forward pass, it is strange that there is a relationship introduce between them by the inverse.

5:35pm. I never thought about it like that. There is likely stuff I do not understand about linear algebra after all.

What about inverting multidimensional matrices and not just 2d. How would that work? Well all multidimensional problems can be pared down to the 2d case...

So I guess it makes sense that 1d could also be pared to the 2d case.

7:25pm. I can't believe I could not go through 5 lectures. But two hours ago I've been seized by some powerful inspiration. I understand how I could make a optimization algorithm for NNs that does reversibility in the outer level, but with gradient descent in the inner.

RNNs would not be applicable to this though, so there is definitely room to go. It does resemble EM in some ways, but it is a real supervised ML algorithm apart from that. It meets all my requirements for RL, including granular credit assignment.

7:30pm. It might be worth implementing. It just might be enough to tide me over until the real thing arrives.

I'll continue studying math, but after I am done with Wildberger's lectures I'll get to it. I might go through other courses apart from topology.

7:40pm. Damn, the today did not go to plan. That surge of inspiration drained me to a husk. I meant to stop for the day a lot earlier.

Let me finally unwind."